Have shared-memory Flush+Reload side-channel working from attacker, after triggering on victim process. Stashing progress thus far as it is in a working state.

+++ defines.h
@@ -20,7 +20,7 @@
 
 #pragma pack(push, 1)
 struct msg {
-  size_t x;
+  uint64_t x;
 };
 #pragma pack(pop)
